add read from file

+++ lab_01/src/utils.hpp
@@ -2,8 +2,9 @@
 #define UTILS_HPP__
 
 #include <QString>
+#include "errors.hpp"
 
 const char *convert_QString_to_char(const QString &);
 
-err_t read_int_number(FILE *, int &);
+err_t read_int_number(FILE *, size_t &);
 #endif // UTILS_HPP__
